kenzo: Property Updates
This is a modified commit. Initial commit was by @MOVZX in common as well in kenzo tree seperately. I have removed lots of bits which were not required.
-Includes Codecs Props
-RmNet Data Props
-Enabled MultiAac offload
-Choose between Virtual or External WFD display

I don't know the original authors, so I'm giving authorship to @MOVZX because I got all these props from him and the original commit was by him.  Also, huge thanks to XDA.

Signed-off-by: AmolAmrit <amol.amrit99@gmail.com>